+++ src/test/java/org/motovs/larisa/qrgenerator/BitBufferTests.java
@@ -9,9 +9,9 @@ public class BitBufferTests {
     @Test
     public void testPush() {
         BitBuffer bitBuffer = new BitBuffer();
-        bitBuffer.push((byte) 7, 4);
+        bitBuffer.push(7, 4);
         assertEquals("0111", bitBuffer.toString());
-        bitBuffer.push((byte) 250, 8);
+        bitBuffer.push(250, 8);
         assertEquals("01111111 1010", bitBuffer.toString());
     }
